Fix issue where mandatory values were not written to XML
On XML creation the code was not writing numeric tags if they had a zero value irrespective of whether the tag was mandatory or not. This commit updates the classes impacted to resolve this. The java code generation has not been updated to automatically deal with this so will need updating in a later commit