Improve check for tag differences from BtS OOB
The original check only took into account whether a child tag existed in the BtS schema and not whether it existed as a child of the specific tag in question. This commit fixes that so that a tag will be identified as changed from BtS if its child tags also differ.