Handle help generation where schemas have different tag definitions for the same tag name
There are cases where schema files use compound tags with the same name for infos, but have different definitions of the child elements. Previously the schema parser would store the definition of each tag based on the tag name so the definition from later schemas parsed would overwrite the definition of all previous schemas. In order to resolve this the folder containing the schema is prepended to the tag name to allow each schema file to have it's own definition. This works because there is only ever a single schema file in each folder in Civ4.