Initial code allowing custom custom tag code to be processed in auto-generated code
Some infos have tags that cannot be processed using the code auto-generator and required hand crafted XML adapters. These infos required that any new tags were manually added otherwise some of the custom code would be overwritten. This commit allows the custom code that would have been overwritten to be defined and incorporated as part of the auto-generation process. This means that future standard tags need not be hand crafted.