Added template class to simplify resource array tag processing
There are multiple locations where info tags contain lists where a resource modifies another list. An example of this is where in buildings specialists can change yields. The tags are in the following form:

<SpecialistYieldChanges>
  <SpecialistYieldChange>
    <SpecialistType></SpecialistType>
    <YieldChanges>
      <iYield>0</iYield>
    </YieldChanges>
  </SpecialistYieldChange>
</SpecialistYieldChanges>

This update can create the custom tag and adapter classes along with the normal import/export code by providing the tag values to a template. It would be possible to have the tooling create all of these without requiring the use of a template, but the effort to try and identify the scenario is not considered worth the additional effort.