Moved InfoProcessor classes to their own package
These classes are dedicated to specific info files and have been moved to their own package for clarity. The main .java package should only contain code that is common and need not be changed unless additional functionality is required. The InfoProcessor files may require updates with changes to the game XML schema.