Made TagInstance the single source of a tag data type
TagProcessors were responsible for storing the tag datatype but this led to problems where in some cases the datatype should remain as it was. The tag processor now directly updates the TagInstance object if the datatype needs to change. This simplifies much of the other code as there is no need to check for the existence of a custom processor before using the tag datatype value.

The population of the code generator data has now been moved to before any custom processors are initialised to allow them to retrieve the TagInstance object if required.